Final
Health bar system refuses to work properly do I made sure that the meta UI reflected it better. added an in game menu/option screen for people to interact with. minor design choices to make it look pretty. mainly focused on trying to make the UI work and not the map or other features